+++ index.js
@@ -11,7 +11,7 @@ const security = require('./lib/security');
 const otps = require('./lib/rave.otps');
 const payment_plan = require('./lib/rave.payment_plan');
 const settlement = require('./lib/rave.settlements');
-const subaccount = require('./lib/rave.subaccount');
+const subaccount_collection = require('./lib/rave.subaccount_collection');
 const subscription = require('./lib/rave.subscriptions');
 const tokenized = require('./lib/rave.tokenized');
 const transaction = require('./lib/rave.transactions');
@@ -39,7 +39,7 @@ const Rave = function (public_key, public_secret, base_url_or_production_flag) {
   this.PaymentPlan = new payment_plan(ravebase);
   this.Settlement = new settlement(ravebase);
   this.Subscription = new subscription(ravebase);
-  this.Subaccount = new subaccount(ravebase);
+  this.CollectionSubaccount = new subaccount_collection(ravebase);
   this.Tokenized = new tokenized(ravebase);
   this.Transaction = new transaction(ravebase);
   this.Transfer = new transfer(ravebase);
