+++ services/subaccount-collection/rave.create.js
@@ -1,11 +1,11 @@
 const { logger } = require('../../utils/logger');
 const { validator } = require('../../utils/validator');
-const { subaccountSchema } = require('../schema/create');
+const { collectionSubaccountSchema } = require('../schema/create');
 
 async function service(data, _rave) {
-  validator(subaccountSchema, data);
+  validator(collectionSubaccountSchema, data);
   const { body: response } = await _rave.request(`v3/subaccounts`, data);
-  logger(`Create a subaccount`, _rave);
+  logger(`Create a collection subaccount`, _rave);
   return response;
 }
