+++ services/subaccount-collection/rave.delete.js
@@ -1,15 +1,15 @@
 const { logger } = require('../../utils/logger');
 const { validator } = require('../../utils/validator');
-const { listSchema } = require('../schema/base');
+const { fetchSchema } = require('../schema/base');
 
 async function service(data, _rave) {
-  validator(listSchema, data);
-  data.method = 'PUT';
+  validator(fetchSchema, data);
+  data.method = 'DELETE';
   const { body: response } = await _rave.request(
     `/v3/subaccounts/${data.id}`,
     data,
   );
-  logger(`update subaccount details`, _rave);
+  logger(`Delete a collection subaccount`, _rave);
   return response;
 }
