+++ services/subaccount-collection/rave.fetch.all.js
@@ -3,7 +3,7 @@ const { handleEmptyFetch } = require('../../utils/build');
 async function service(data, _rave) {
   return handleEmptyFetch(
     data,
-    `Fetch all subaccount`,
+    `Fetch all collection subaccounts`,
     `/v3/subaccounts?`,
     _rave,
   );
